+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/extend/nacos/NacosContextPropertiesTest.java
@@ -34,7 +34,14 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.util.CollectionUtils;
 
+import static com.tencent.polaris.plugins.connector.common.constant.NacosConstant.MetadataMapKey.NACOS_CLUSTER_KEY;
+import static com.tencent.polaris.plugins.connector.common.constant.NacosConstant.MetadataMapKey.NACOS_GROUP_KEY;
 import static org.assertj.core.api.Assertions.assertThat;
+import static shade.polaris.com.alibaba.nacos.api.PropertyKeyConst.CONTEXT_PATH;
+import static shade.polaris.com.alibaba.nacos.api.PropertyKeyConst.NAMESPACE;
+import static shade.polaris.com.alibaba.nacos.api.PropertyKeyConst.PASSWORD;
+import static shade.polaris.com.alibaba.nacos.api.PropertyKeyConst.USERNAME;
+
 
 /**
  * Test for {@link NacosContextProperties}.
@@ -55,7 +62,6 @@ public class NacosContextPropertiesTest {
 	@Test
 	public void testDefaultInitialization() {
 		assertThat(nacosContextProperties).isNotNull();
-		assertThat(nacosContextProperties.isEnabled()).isTrue();
 		assertThat(nacosContextProperties.getServerAddr()).isEqualTo("127.0.0.1:8848");
 		assertThat(nacosContextProperties.isRegisterEnabled()).isTrue();
 		assertThat(nacosContextProperties.isDiscoveryEnabled()).isTrue();
@@ -82,11 +88,12 @@ public class NacosContextPropertiesTest {
 		assertThat(DefaultPlugins.SERVER_CONNECTOR_NACOS.equals(serverConnectorConfig.getProtocol())).isTrue();
 
 		Map<String, String> metadata = serverConnectorConfig.getMetadata();
-		assertThat(metadata.get(NacosConfigModifier.USERNAME)).isEqualTo(nacosContextProperties.getUsername());
-		assertThat(metadata.get(NacosConfigModifier.PASSWORD)).isEqualTo(nacosContextProperties.getPassword());
-		assertThat(metadata.get(NacosConfigModifier.CONTEXT_PATH)).isEqualTo(nacosContextProperties.getContextPath());
-		assertThat(metadata.get(NacosConfigModifier.NAMESPACE)).isEqualTo(nacosContextProperties.getNamespace());
-		assertThat(metadata.get(NacosConfigModifier.GROUP)).isEqualTo(nacosContextProperties.getGroup());
+		assertThat(metadata.get(USERNAME)).isEqualTo(nacosContextProperties.getUsername());
+		assertThat(metadata.get(PASSWORD)).isEqualTo(nacosContextProperties.getPassword());
+		assertThat(metadata.get(CONTEXT_PATH)).isEqualTo(nacosContextProperties.getContextPath());
+		assertThat(metadata.get(NAMESPACE)).isEqualTo(nacosContextProperties.getNamespace());
+		assertThat(metadata.get(NACOS_GROUP_KEY)).isEqualTo(nacosContextProperties.getGroup());
+		assertThat(metadata.get(NACOS_CLUSTER_KEY)).isEqualTo(nacosContextProperties.getClusterName());
 	}
 
 	@SpringBootApplication