+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - [docs:optimize tsf example.](https://github.com/Tencent/spring-cloud-tencent/pull/1710)
 - [feat:support TSF certificate manager.](https://github.com/Tencent/spring-cloud-tencent/pull/1715)
 - [feat:support tsf unit.](https://github.com/Tencent/spring-cloud-tencent/pull/1681)
+- [feat: support service registry and discovery with Polaris and Nacos](https://github.com/Tencent/spring-cloud-tencent/pull/1718)